Move logic from Fragment to ViewModel
