+++ app/src/main/java/io/github/droidkaigi/confsched2017/viewmodel/SponsorshipsViewModel.java
@@ -20,31 +20,46 @@ import io.github.droidkaigi.confsched2017.R;
 import io.github.droidkaigi.confsched2017.model.Sponsorship;
 import io.github.droidkaigi.confsched2017.view.helper.ResourceResolver;
 import io.reactivex.Single;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.disposables.Disposable;
+import timber.log.Timber;
 
 public final class SponsorshipsViewModel extends BaseObservable implements ViewModel {
+
+    private static final String TAG = SponsorshipsViewModel.class.getSimpleName();
 
     private ResourceResolver resourceResolver;
 
     private final ObservableList<SponsorshipViewModel> sponsorshipViewModels;
 
+    private final CompositeDisposable compositeDisposable;
+
     @Inject
-    SponsorshipsViewModel(ResourceResolver resourceResolver) {
+    SponsorshipsViewModel(ResourceResolver resourceResolver, CompositeDisposable compositeDisposable) {
         this.resourceResolver = resourceResolver;
+        this.compositeDisposable = compositeDisposable;
         this.sponsorshipViewModels = new ObservableArrayList<>();
+    }
+
+    public void start() {
+        Disposable disposable = loadSponsors()
+                .map(this::convertToViewModel)
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(
+                        this::renderSponsorships,
+                        throwable -> Timber.tag(TAG).e(throwable, "Failed to show sponsors.")
+                );
+        compositeDisposable.add(disposable);
     }
 
     @Override
     public void destroy() {
-        // No-op
+        compositeDisposable.clear();
     }
 
     public ObservableList<SponsorshipViewModel> getSponsorShipViewModels() {
         return sponsorshipViewModels;
-    }
-
-    public Single<List<SponsorshipViewModel>> convertSponsorShip() {
-        return loadSponsors().map(sponsorships ->
-                Stream.of(sponsorships).map(SponsorshipViewModel::new).collect(Collectors.toList()));
     }
 
     private Single<List<Sponsorship>> loadSponsors() {
@@ -66,5 +81,14 @@ public final class SponsorshipsViewModel extends BaseObservable implements ViewM
         final Type listType = new TypeToken<List<Sponsorship>>() {
         }.getType();
         return gson.fromJson(json, listType);
+    }
+
+    private List<SponsorshipViewModel> convertToViewModel(List<Sponsorship> sponsorships) {
+        return Stream.of(sponsorships).map(SponsorshipViewModel::new).collect(Collectors.toList());
+    }
+
+    private void renderSponsorships(List<SponsorshipViewModel> sponsorships) {
+        sponsorshipViewModels.clear();
+        sponsorshipViewModels.addAll(sponsorships);
     }
 }
